requirements.txt
aiofiles==23.1.0
aiogram==3.0.0rc1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2023.7.22
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pydantic==2.1.1
pydantic-settings==2.0.3
pydantic_core==2.4.0
python-dotenv==1.0.0
typing_extensions==4.7.1
yarl==1.8.2
psycopg2-binary==2.9.9

requests~=2.31.0